Refactor Favorites and Recipes Fragments with Database Updates
-   Updated `FavoritesFragment` to fetch and display favorite meals with details from the database.
-   Implemented error handling and logging in `FavoritesFragment`.
-   Refactored `RecipesFragment` to fetch recipes by category concurrently.
-   Improved database interactions in `RecipesAdapter`.
-   Modified database schema to version 4.
-   Corrected meal instruction storage in the database.
-   Add log on the `RecipesFragment`
-   Add default values on `Meal` class
-   Add `FavoriteMeal` to the update functions
-   Update the title of the `FavoritesFragment` to "Favorites"

+++ app/src/main/java/com/maayn/mealmate/presentation/home/HomeViewModel.kt
@@ -89,7 +89,7 @@ class HomeViewModel(application: Application) : AndroidViewModel(application) {
                     IngredientEntity(mealId = mealEntity.id, name = it.name, measure = it.measure)
                 }
                 val instructions = apiMeal.extractInstructions().map {
-                    InstructionEntity(mealId = mealEntity.id, step = it.step.toInt(), description = it.step)
+                    InstructionEntity(mealId = mealEntity.id, step = it.step, description = it.step)
                 }
 
                 // 🔹 Store data in Firestore and Room
